+++ Projects/App/Sources/ContentView.swift
@@ -11,9 +11,6 @@ import DesignSystem
 
 struct ContentView: View {
     var body: some View {
-//        RouterView {
-//            TabbarMainView()
-//        }
         Text("ContentView")
     }
 }
